requirements.txt
click
pyyaml
requests
jinja2